refactor: Consolidate framework and control views into unified structure
Restructure framework and control views to improve code organization:
- Move framework-related components from `/frameworks` to `/overview/frameworks`
- Relocate control-specific components to `/overview/frameworks/controls`
- Remove duplicate action, hook, and component files
- Update file paths and imports to reflect new structure
- Standardize component naming and file organization
- Improve code reusability and maintainability
