+++ apps/app/src/app/[locale]/(app)/(dashboard)/(home)/layout.tsx
@@ -2,17 +2,17 @@ import { getI18n } from "@/locales/server";
 import { SecondaryMenu } from "@bubba/ui/secondary-menu";
 
 export default async function Layout({
-  children,
+	children,
 }: {
-  children: React.ReactNode;
+	children: React.ReactNode;
 }) {
-  const t = await getI18n();
+	const t = await getI18n();
 
-  return (
-    <div className="max-w-[1200px] m-auto">
-      <SecondaryMenu items={[{ path: "/", label: t("overview.title") }]} />
+	return (
+		<div className="max-w-[1200px] mx-auto">
+			<SecondaryMenu items={[{ path: "/", label: t("overview.title") }]} />
 
-      <main className="mt-8">{children}</main>
-    </div>
-  );
+			<main className="mt-8">{children}</main>
+		</div>
+	);
 }
