+++ apps/app/src/app/[locale]/(app)/(dashboard)/(home)/overview/frameworks/controls/[id]/actions/getOrganizationControlRequirements.ts
