+++ apps/app/src/locales/en.ts
@@ -953,7 +953,7 @@ export default {
 		title: "Evidence",
 		list: "All Evidence",
 		overview: "Evidence Overview",
-		edit: "Edit Evidence",
+		edit: "Uploaded Evidence",
 		dashboard: {
 			layout: "Dashboard",
 			layout_back_button: "Back",
